+++ api/src/shared/infrastructure/repositories/country-repository.js
@@ -1,5 +1,5 @@
-import { knex } from '../../../../../db/knex-database-connection.js';
-import { Country } from '../../../../shared/domain/read-models/Country.js';
+import { knex } from '../../../../db/knex-database-connection.js';
+import { Country } from '../../domain/read-models/Country.js';
 
 const findAll = async function () {
   const result = await knex
